+++ src/test/java/s32a/airhockey/GameTest.java
@@ -42,9 +42,9 @@ public class GameTest
     public void setUp()
     {
         //Player starter
-        starter = new Player("testPlayer", 0, Colors.Red);
+        starter = new Player("testPlayer", (double)0, Colors.Red);
         game = new Game(starter);
-        spec = new Spectator("testSpectator", 0, game);
+        spec = new Spectator("testSpectator", (double)0, game);
         game.addSpectator(spec);
     }
     
@@ -101,9 +101,9 @@ public class GameTest
     @Test
     public void testAddPlayerGameFull()
     {
-        Player p2 = new Player("testPlayer2", 0, Colors.Blue);
-        Player p3 = new Player("testPlayer3", 0, Colors.Green);
-        Player p4 = new Player("testPlayer4", 0, Colors.Green);
+        Player p2 = new Player("testPlayer2", (double)0, Colors.Blue);
+        Player p3 = new Player("testPlayer3", (double)0, Colors.Green);
+        Player p4 = new Player("testPlayer4", (double)0, Colors.Green);
         
         game.addPlayer(p2);
         game.addPlayer(p3);
@@ -146,7 +146,7 @@ public class GameTest
     @Test
     public void testRemoveSpectatorNotAParticipant()
     {
-        Spectator spec2 = new Spectator("testSpectator2", 0, game);
+        Spectator spec2 = new Spectator("testSpectator2", (double)0, game);
         
         Boolean expResult = false;
         Boolean result = game.removeSpectator(spec2);
@@ -169,8 +169,8 @@ public class GameTest
     @Test
     public void testBeginGameAlreadyBegon()
     {
-        Player p2 = new Player("testPlayer2", 0, Colors.Blue);
-        Player p3 = new Player("testPlayer3", 0, Colors.Green);
+        Player p2 = new Player("testPlayer2", (double)0, Colors.Blue);
+        Player p3 = new Player("testPlayer3", (double)0, Colors.Green);
         
         game.addPlayer(p2);
         game.addPlayer(p3);
@@ -255,7 +255,7 @@ public class GameTest
     @Test
     public void testGetNextColorGreen()
     {
-        game.addPlayer(new Player("testPlayer2", 0, Colors.Blue));
+        game.addPlayer(new Player("testPlayer2", (double)0, Colors.Blue));
         
         Colors expResult = Colors.Green;
         Colors result = game.getNextColor();
